phase2 completed not debugged

+++ discount.cpp
@@ -23,21 +23,21 @@ specific_food_discount::specific_food_discount()
 void total_price_discount::set_properties(std::string type_, std::string min, std::string value_)
 {
     type = type_;
-    minimum_total = std::stoi(min);
-    value = std::stoi(value_);
+    minimum_total = std::stof(min);
+    value = std::stof(value_);
 }
 
 void first_order_discount::set_properties(std::string type_, std::string value_, std::string)
 {
     type = type_;
-    value = std::stoi(value_);
+    value = std::stof(value_);
 }
 
 void specific_food_discount::set_properties(std::string type_, std::string food_name_, std::string value_)
 {
     type = type_;
     food_name = food_name_;
-    value = std::stoi(value_);
+    value = std::stof(value_);
 }
 
 std::string specific_food_discount::get_food_name()
